feat: add SelectVisibility component and integrate into EditNote
